requirements.txt
PyQt5
pyOpenSSL
Jinja2